refactor cli code from pkg to cmd (#3591)
* refactor cli code from pkg to cmd


* fixes in imports


* fixes tests


* fixed conflicts


* moved non-commands to utils
