+++ cmd/cli/kubectl-kyverno/apply/generate.go
@@ -4,8 +4,8 @@ import (
 	"reflect"
 
 	report "github.com/kyverno/kyverno/api/policyreport/v1alpha2"
+	sanitizederror "github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/utils/sanitizedError"
 	client "github.com/kyverno/kyverno/pkg/dclient"
-	sanitizederror "github.com/kyverno/kyverno/pkg/kyverno/sanitizedError"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"sigs.k8s.io/controller-runtime/pkg/log"
