+++ cmd/cli/kubectl-kyverno/apply/generate_test.go
