+++ cmd/cli/kubectl-kyverno/utils/common/common.go
@@ -19,6 +19,8 @@ import (
 	v1 "github.com/kyverno/kyverno/api/kyverno/v1"
 	v1beta1 "github.com/kyverno/kyverno/api/kyverno/v1beta1"
 	report "github.com/kyverno/kyverno/api/policyreport/v1alpha2"
+	sanitizederror "github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/utils/sanitizedError"
+	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/utils/store"
 	"github.com/kyverno/kyverno/pkg/autogen"
 	client "github.com/kyverno/kyverno/pkg/dclient"
 	"github.com/kyverno/kyverno/pkg/engine"
@@ -26,8 +28,6 @@ import (
 	"github.com/kyverno/kyverno/pkg/engine/response"
 	ut "github.com/kyverno/kyverno/pkg/engine/utils"
 	"github.com/kyverno/kyverno/pkg/engine/variables"
-	sanitizederror "github.com/kyverno/kyverno/pkg/kyverno/sanitizedError"
-	"github.com/kyverno/kyverno/pkg/kyverno/store"
 	"github.com/kyverno/kyverno/pkg/policymutation"
 	"github.com/kyverno/kyverno/pkg/policyreport"
 	"github.com/kyverno/kyverno/pkg/utils"
